chore(custom-server): move and upgrade to TypeScript
changing the initial location for the custom-server files running under-the-hood for sveltekit into
`server/`
